get-gpg-key: use a more resilient key fetching
Ubuntu's GPG servers sometimes threw errors if presented with TLSv1.3
capable handshake messages.

Henceforth keyservers from the SKS pool will be used. This spreads
load when images are built at the same time (among pool members), decreases
the probability of a widespread failure when fetching keys (in case
Ubuntu's host is down), and due to a second connection attempt to a
then new host (it's a rotation of hosts when using the pool) individual
resiliency is increased.