update 18.04 for amd64
Updates Perl (it's version bump, really) and pulls in the bash
which is used in 16.04. The latter has not been picked up because
of a versioning change, and Ubuntu's version, albeit at lower
patch level, did overwrite mine.

Size of 18.04: 38.4MiB

Please keep in mind that 18.04 has not been released officially, yet.
This is an unsupported preview.