curl: upgrade to 7.61.1-65 due to errors with http2
http2 has been switched on by default for apt, but some
servers (including mine) run still older implementations.
This in turn resulted in 'apt-get update' and similar to skip those
servers in account of "invalid http header(s)."

http2 is supposed to be the default come the next version of
curl (7.62?), not this one.

In a gist, this brings this setting in line with vanilla curl 7.61.1.